Feature (Oauth Transparent Migration)
What does this PR do?

Make Oauth transparent migration
